some re-factoring... Item -> GetItem, Last -> GetLast; convenience returns to the *this for lists and standardising list interface... the mode grow -v is 'optimised'; the matrix multiplication and inversion caches added to the unit cell; fixes regarding the MatrixCount and EquivCount - the latter is to be used in weights etc!; also tried to get rid of the matrix list in TSAtom (one matrix + the list of equivs of the TCAtom should be enough...), but failed;

+++ branches/dev/glib/gpcollection.h
@@ -31,7 +31,7 @@ public:
   TGlPrimitive& NewPrimitive(const olxstr& Name, short type);
   void AddPrimitive(TGlPrimitive& P)  {  Primitives.Add(P);  }
 
-  void RemovePrimitive(int index) {  Primitives.Delete(index); }
+  void RemovePrimitive(size_t index) {  Primitives.Delete(index); }
   void RemovePrimitive(TGlPrimitive& GlP)  {  Primitives.Remove(GlP);  }
 
   bool ContainsPrimitive(TGlPrimitive& GlP);
